requirements.txt
Django==3.2.11
djangorestframework==3.13.1
knox==0.1.14
gunicorn==20.1.0
virtualenv==20.4.7
whitenoise==5.3.0
django-heroku==0.3.1
paramiko==2.10.2
cryptography==36.0.1
Jinja2==3.0.1
requests==2.25.1
boto3==1.21.20
botocore==1.24.20
psycopg2==2.9.1